rename imports in service/controller/azure/v7patch1: github.com/giantswarm/cluster-operator/pkg/v7/key > github.com/giantswarm/cluster-operator/pkg/v7patch1/key
