+++ service/controller/azure/v7patch1/key/key.go
@@ -5,7 +5,7 @@ import (
 	"github.com/giantswarm/microerror"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/cluster-operator/pkg/v7/key"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/key"
 )
 
 // ChartSpecs returns charts installed only for Azure.
