+++ service/controller/azure/v7patch1/resource/chartconfig/current.go
@@ -9,7 +9,7 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
 	"github.com/giantswarm/cluster-operator/pkg/v7/chartconfig"
-	"github.com/giantswarm/cluster-operator/pkg/v7/key"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/key"
 	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v7/key"
 )
