+++ service/controller/azure/v7patch1/resource/configmap/update.go
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/operatorkit/controller/context/reconciliationcanceledcontext"
 
 	"github.com/giantswarm/cluster-operator/pkg/v7/configmap"
-	"github.com/giantswarm/cluster-operator/pkg/v7/key"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/key"
 	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v7/key"
 )
